apusys power: fix warning of cfi checking
[Detail]
WARNING: CPU: 4 PID: 151 at kernel/cfi.c:30 __ubsan_handle_cfi_check_fail+0x4c/0x54
[  294.196964] pc : __ubsan_handle_cfi_check_fail+0x4c/0x54
[  294.197630] lr : __ubsan_handle_cfi_check_fail+0x4c/0x54
[  294.208706] Call trace:
[  294.209014]  __ubsan_handle_cfi_check_fail+0x4c/0x54
[  294.209686]  __cfi_check_fail+0x30/0x38 [apusys]
[  294.210315]  __cfi_check+0x1b18/0x1bc0 [apusys]
[  294.210897]  apupw_dbg_dvfs_tag_update+0x100/0x150 [apu_top]
[  294.211618]  ausr_get_target_freq+0xc4/0x14c [apu_top]
[  294.212267]  update_devfreq+0x64/0x174
[  294.212748]  apu_qos_set_vcore+0xc8/0x160 [apu_top]
[  294.213410]  qos_work_func+0x238/0x394 [apusys]
[  294.213981]  process_one_work+0x2d8/0x580
[  294.214485]  worker_thread+0x28c/0x518
[  294.214957]  kthread+0x174/0x184
[  294.215364]  ret_from_fork+0x10/0x18

(cherry-picked from commit 82ba3ba2d21ccbf13c43a0d9ee43d08f60dd3a23)
CR-Id: ALPS05516247
Feature: [Module] APUSYS

+++ drivers/misc/mediatek/apusys/power/2.5/inc/events/apu_events.h
@@ -103,8 +103,8 @@ TRACE_EVENT(apupwr_dvfs,
 		__string(gov_name, gov_name)
 		__string(p_name, p_name)
 		__string(c_name, c_name)
-		__field(unsigned int, opp)
-		__field(unsigned int, freq)
+		__field(u32, opp)
+		__field(ulong, freq)
 	),
 	TP_fast_assign(
 		__assign_str(gov_name, gov_name);
